fix(wif): require the config in the contract and validate at its owning layer
Two review findings on the pushed branch:

- Mark WorkloadIdentity.workload_identity_config REQUIRED. The server
  rejects a Create or a config-mask Update that omits it, but the proto
  left the field optional, so the generated contract disagreed with the
  handler. Only the WorkloadIdentity descriptor and its OpenAPI required
  list change; the rest of the OpenAPI regeneration is unpinned-plugin
  drift and is left out.

- Move TestWorkloadIdentityConfigValidation from backend/tests to
  backend/api/v1. It pins that Create and Update reach the validator,
  which needs no server boot; it now calls the handlers on a metadata
  fixture. AGENTS.md reserves a full server for a runner, rollout, or
  audit trail.
