+++ frontend/src/types/proto-es/v1/workload_identity_service_pb.js
@@ -16,7 +16,7 @@ import { file_v1_common } from "./common_pb";
  * Describes the file v1/workload_identity_service.proto.
  */
 export const file_v1_workload_identity_service = /*@__PURE__*/
-  fileDesc("CiJ2MS93b3JrbG9hZF9pZGVudGl0eV9zZXJ2aWNlLnByb3RvEgtieXRlYmFzZS52MSKyAgoQV29ya2xvYWRJZGVudGl0eRIRCgRuYW1lGAEgASgJQgPgQQMSIQoFc3RhdGUYAiABKA4yEi5ieXRlYmFzZS52MS5TdGF0ZRISCgVlbWFpbBgDIAEoCUID4EEDEhcKBXRpdGxlGAQgASgJQgi6SAVyAxjIARI0CgtjcmVhdGVfdGltZRgFIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXBCA+BBAxJFChh3b3JrbG9hZF9pZGVudGl0eV9jb25maWcYBiABKAsyIy5ieXRlYmFzZS52MS5Xb3JrbG9hZElkZW50aXR5Q29uZmlnOj7qQTsKHWJ5dGViYXNlLmNvbS9Xb3JrbG9hZElkZW50aXR5Ehp3b3JrbG9hZElkZW50aXRpZXMve2VtYWlsfSKgAgoWV29ya2xvYWRJZGVudGl0eUNvbmZpZxJMCg1wcm92aWRlcl90eXBlGAEgASgOMjAuYnl0ZWJhc2UudjEuV29ya2xvYWRJZGVudGl0eUNvbmZpZy5Qcm92aWRlclR5cGVCA+BBAhIXCgppc3N1ZXJfdXJsGAIgASgJQgPgQQISHgoRYWxsb3dlZF9hdWRpZW5jZXMYAyADKAlCA+BBAhIcCg9zdWJqZWN0X3BhdHRlcm4YBCABKAlCA+BBAhIQCghqd2tzX3VybBgFIAEoCSJPCgxQcm92aWRlclR5cGUSHQoZUFJPVklERVJfVFlQRV9VTlNQRUNJRklFRBAAEgoKBkdJVEhVQhABEgoKBkdJVExBQhACEggKBE9JREMQAyKsAQodQ3JlYXRlV29ya2xvYWRJZGVudGl0eVJlcXVlc3QSKQoGcGFyZW50GAEgASgJQhn6QRYKFGJ5dGViYXNlLmNvbS9Qcm9qZWN0EiEKFHdvcmtsb2FkX2lkZW50aXR5X2lkGAIgASgJQgPgQQISPQoRd29ya2xvYWRfaWRlbnRpdHkYAyABKAsyHS5ieXRlYmFzZS52MS5Xb3JrbG9hZElkZW50aXR5QgPgQQIiUQoaR2V0V29ya2xvYWRJZGVudGl0eVJlcXVlc3QSMwoEbmFtZRgBIAEoCUIl4EEC+kEfCh1ieXRlYmFzZS5jb20vV29ya2xvYWRJZGVudGl0eSKXAQodTGlzdFdvcmtsb2FkSWRlbnRpdGllc1JlcXVlc3QSKQoGcGFyZW50GAEgASgJQhn6QRYKFGJ5dGViYXNlLmNvbS9Qcm9qZWN0EhEKCXBhZ2Vfc2l6ZRgCIAEoBRISCgpwYWdlX3Rva2VuGAMgASgJEhQKDHNob3dfZGVsZXRlZBgEIAEoCBIOCgZmaWx0ZXIYBSABKAkidQoeTGlzdFdvcmtsb2FkSWRlbnRpdGllc1Jlc3BvbnNlEjoKE3dvcmtsb2FkX2lkZW50aXRpZXMYASADKAsyHS5ieXRlYmFzZS52MS5Xb3JrbG9hZElkZW50aXR5EhcKD25leHRfcGFnZV90b2tlbhgCIAEoCSKPAQodVXBkYXRlV29ya2xvYWRJZGVudGl0eVJlcXVlc3QSPQoRd29ya2xvYWRfaWRlbnRpdHkYASABKAsyHS5ieXRlYmFzZS52MS5Xb3JrbG9hZElkZW50aXR5QgPgQQISLwoLdXBkYXRlX21hc2sYAiABKAsyGi5nb29nbGUucHJvdG9idWYuRmllbGRNYXNrIlQKHURlbGV0ZVdvcmtsb2FkSWRlbnRpdHlSZXF1ZXN0EjMKBG5hbWUYASABKAlCJeBBAvpBHwodYnl0ZWJhc2UuY29tL1dvcmtsb2FkSWRlbnRpdHkiVgofVW5kZWxldGVXb3JrbG9hZElkZW50aXR5UmVxdWVzdBIzCgRuYW1lGAEgASgJQiXgQQL6QR8KHWJ5dGViYXNlLmNvbS9Xb3JrbG9hZElkZW50aXR5MsULChdXb3JrbG9hZElkZW50aXR5U2VydmljZRLPAgoWQ3JlYXRlV29ya2xvYWRJZGVudGl0eRIqLmJ5dGViYXNlLnYxLkNyZWF0ZVdvcmtsb2FkSWRlbnRpdHlSZXF1ZXN0Gh0uYnl0ZWJhc2UudjEuV29ya2xvYWRJZGVudGl0eSLpAdpBLXBhcmVudCx3b3JrbG9hZF9pZGVudGl0eSx3b3JrbG9hZF9pZGVudGl0eV9pZIrqMBxiYi53b3JrbG9hZElkZW50aXRpZXMuY3JlYXRlkOowAZjqMAGg6jADqOowBoLT5JMCggE6EXdvcmtsb2FkX2lkZW50aXR5WkE6EXdvcmtsb2FkX2lkZW50aXR5IiwvdjEve3BhcmVudD13b3Jrc3BhY2VzLyp9L3dvcmtsb2FkSWRlbnRpdGllcyIqL3YxL3twYXJlbnQ9cHJvamVjdHMvKn0vd29ya2xvYWRJZGVudGl0aWVzErYBChNHZXRXb3JrbG9hZElkZW50aXR5EicuYnl0ZWJhc2UudjEuR2V0V29ya2xvYWRJZGVudGl0eVJlcXVlc3QaHS5ieXRlYmFzZS52MS5Xb3JrbG9hZElkZW50aXR5IlfaQQRuYW1liuowGWJiLndvcmtsb2FkSWRlbnRpdGllcy5nZXSQ6jABoOowBKjqMAmC0+STAiESHy92MS97bmFtZT13b3JrbG9hZElkZW50aXRpZXMvKn0SiQIKFkxpc3RXb3JrbG9hZElkZW50aXRpZXMSKi5ieXRlYmFzZS52MS5MaXN0V29ya2xvYWRJZGVudGl0aWVzUmVxdWVzdBorLmJ5dGViYXNlLnYxLkxpc3RXb3JrbG9hZElkZW50aXRpZXNSZXNwb25zZSKVAdpBBnBhcmVudIrqMBpiYi53b3JrbG9hZElkZW50aXRpZXMubGlzdJDqMAGg6jAEqOowCYLT5JMCXFouEiwvdjEve3BhcmVudD13b3Jrc3BhY2VzLyp9L3dvcmtsb2FkSWRlbnRpdGllcxIqL3YxL3twYXJlbnQ9cHJvamVjdHMvKn0vd29ya2xvYWRJZGVudGl0aWVzEoICChZVcGRhdGVXb3JrbG9hZElkZW50aXR5EiouYnl0ZWJhc2UudjEuVXBkYXRlV29ya2xvYWRJZGVudGl0eVJlcXVlc3QaHS5ieXRlYmFzZS52MS5Xb3JrbG9hZElkZW50aXR5IpwB2kEdd29ya2xvYWRfaWRlbnRpdHksdXBkYXRlX21hc2uK6jAcYmIud29ya2xvYWRJZGVudGl0aWVzLnVwZGF0ZZDqMAGY6jABoOowA6jqMAaC0+STAkY6EXdvcmtsb2FkX2lkZW50aXR5MjEvdjEve3dvcmtsb2FkX2lkZW50aXR5Lm5hbWU9d29ya2xvYWRJZGVudGl0aWVzLyp9ErwBChZEZWxldGVXb3JrbG9hZElkZW50aXR5EiouYnl0ZWJhc2UudjEuRGVsZXRlV29ya2xvYWRJZGVudGl0eVJlcXVlc3QaFi5nb29nbGUucHJvdG9idWYuRW1wdHkiXtpBBG5hbWWK6jAcYmIud29ya2xvYWRJZGVudGl0aWVzLmRlbGV0ZZDqMAGY6jABoOowBKjqMAmC0+STAiEqHy92MS97bmFtZT13b3JrbG9hZElkZW50aXRpZXMvKn0SzgEKGFVuZGVsZXRlV29ya2xvYWRJZGVudGl0eRIsLmJ5dGViYXNlLnYxLlVuZGVsZXRlV29ya2xvYWRJZGVudGl0eVJlcXVlc3QaHS5ieXRlYmFzZS52MS5Xb3JrbG9hZElkZW50aXR5ImWK6jAeYmIud29ya2xvYWRJZGVudGl0aWVzLnVuZGVsZXRlkOowAZjqMAGg6jAEqOowCYLT5JMCLToBKiIoL3YxL3tuYW1lPXdvcmtsb2FkSWRlbnRpdGllcy8qfTp1bmRlbGV0ZUKyAQoPY29tLmJ5dGViYXNlLnYxQhxXb3JrbG9hZElkZW50aXR5U2VydmljZVByb3RvUAFaNGdpdGh1Yi5jb20vYnl0ZWJhc2UvYnl0ZWJhc2UvYmFja2VuZC9nZW5lcmF0ZWQtZ28vdjGiAgNCWFiqAgtCeXRlYmFzZS5WMcoCC0J5dGViYXNlXFYx4gIXQnl0ZWJhc2VcVjFcR1BCTWV0YWRhdGHqAgxCeXRlYmFzZTo6VjFiBnByb3RvMw", [file_buf_validate_validate, file_google_api_annotations, file_google_api_client, file_google_api_field_behavior, file_google_api_resource, file_google_protobuf_empty, file_google_protobuf_field_mask, file_google_protobuf_timestamp, file_v1_annotation, file_v1_common]);
+  fileDesc("CiJ2MS93b3JrbG9hZF9pZGVudGl0eV9zZXJ2aWNlLnByb3RvEgtieXRlYmFzZS52MSK3AgoQV29ya2xvYWRJZGVudGl0eRIRCgRuYW1lGAEgASgJQgPgQQMSIQoFc3RhdGUYAiABKA4yEi5ieXRlYmFzZS52MS5TdGF0ZRISCgVlbWFpbBgDIAEoCUID4EEDEhcKBXRpdGxlGAQgASgJQgi6SAVyAxjIARI0CgtjcmVhdGVfdGltZRgFIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXBCA+BBAxJKChh3b3JrbG9hZF9pZGVudGl0eV9jb25maWcYBiABKAsyIy5ieXRlYmFzZS52MS5Xb3JrbG9hZElkZW50aXR5Q29uZmlnQgPgQQI6PupBOwodYnl0ZWJhc2UuY29tL1dvcmtsb2FkSWRlbnRpdHkSGndvcmtsb2FkSWRlbnRpdGllcy97ZW1haWx9IqACChZXb3JrbG9hZElkZW50aXR5Q29uZmlnEkwKDXByb3ZpZGVyX3R5cGUYASABKA4yMC5ieXRlYmFzZS52MS5Xb3JrbG9hZElkZW50aXR5Q29uZmlnLlByb3ZpZGVyVHlwZUID4EECEhcKCmlzc3Vlcl91cmwYAiABKAlCA+BBAhIeChFhbGxvd2VkX2F1ZGllbmNlcxgDIAMoCUID4EECEhwKD3N1YmplY3RfcGF0dGVybhgEIAEoCUID4EECEhAKCGp3a3NfdXJsGAUgASgJIk8KDFByb3ZpZGVyVHlwZRIdChlQUk9WSURFUl9UWVBFX1VOU1BFQ0lGSUVEEAASCgoGR0lUSFVCEAESCgoGR0lUTEFCEAISCAoET0lEQxADIqwBCh1DcmVhdGVXb3JrbG9hZElkZW50aXR5UmVxdWVzdBIpCgZwYXJlbnQYASABKAlCGfpBFgoUYnl0ZWJhc2UuY29tL1Byb2plY3QSIQoUd29ya2xvYWRfaWRlbnRpdHlfaWQYAiABKAlCA+BBAhI9ChF3b3JrbG9hZF9pZGVudGl0eRgDIAEoCzIdLmJ5dGViYXNlLnYxLldvcmtsb2FkSWRlbnRpdHlCA+BBAiJRChpHZXRXb3JrbG9hZElkZW50aXR5UmVxdWVzdBIzCgRuYW1lGAEgASgJQiXgQQL6QR8KHWJ5dGViYXNlLmNvbS9Xb3JrbG9hZElkZW50aXR5IpcBCh1MaXN0V29ya2xvYWRJZGVudGl0aWVzUmVxdWVzdBIpCgZwYXJlbnQYASABKAlCGfpBFgoUYnl0ZWJhc2UuY29tL1Byb2plY3QSEQoJcGFnZV9zaXplGAIgASgFEhIKCnBhZ2VfdG9rZW4YAyABKAkSFAoMc2hvd19kZWxldGVkGAQgASgIEg4KBmZpbHRlchgFIAEoCSJ1Ch5MaXN0V29ya2xvYWRJZGVudGl0aWVzUmVzcG9uc2USOgoTd29ya2xvYWRfaWRlbnRpdGllcxgBIAMoCzIdLmJ5dGViYXNlLnYxLldvcmtsb2FkSWRlbnRpdHkSFwoPbmV4dF9wYWdlX3Rva2VuGAIgASgJIo8BCh1VcGRhdGVXb3JrbG9hZElkZW50aXR5UmVxdWVzdBI9ChF3b3JrbG9hZF9pZGVudGl0eRgBIAEoCzIdLmJ5dGViYXNlLnYxLldvcmtsb2FkSWRlbnRpdHlCA+BBAhIvCgt1cGRhdGVfbWFzaxgCIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5GaWVsZE1hc2siVAodRGVsZXRlV29ya2xvYWRJZGVudGl0eVJlcXVlc3QSMwoEbmFtZRgBIAEoCUIl4EEC+kEfCh1ieXRlYmFzZS5jb20vV29ya2xvYWRJZGVudGl0eSJWCh9VbmRlbGV0ZVdvcmtsb2FkSWRlbnRpdHlSZXF1ZXN0EjMKBG5hbWUYASABKAlCJeBBAvpBHwodYnl0ZWJhc2UuY29tL1dvcmtsb2FkSWRlbnRpdHkyxQsKF1dvcmtsb2FkSWRlbnRpdHlTZXJ2aWNlEs8CChZDcmVhdGVXb3JrbG9hZElkZW50aXR5EiouYnl0ZWJhc2UudjEuQ3JlYXRlV29ya2xvYWRJZGVudGl0eVJlcXVlc3QaHS5ieXRlYmFzZS52MS5Xb3JrbG9hZElkZW50aXR5IukB2kEtcGFyZW50LHdvcmtsb2FkX2lkZW50aXR5LHdvcmtsb2FkX2lkZW50aXR5X2lkiuowHGJiLndvcmtsb2FkSWRlbnRpdGllcy5jcmVhdGWQ6jABmOowAaDqMAOo6jAGgtPkkwKCAToRd29ya2xvYWRfaWRlbnRpdHlaQToRd29ya2xvYWRfaWRlbnRpdHkiLC92MS97cGFyZW50PXdvcmtzcGFjZXMvKn0vd29ya2xvYWRJZGVudGl0aWVzIiovdjEve3BhcmVudD1wcm9qZWN0cy8qfS93b3JrbG9hZElkZW50aXRpZXMStgEKE0dldFdvcmtsb2FkSWRlbnRpdHkSJy5ieXRlYmFzZS52MS5HZXRXb3JrbG9hZElkZW50aXR5UmVxdWVzdBodLmJ5dGViYXNlLnYxLldvcmtsb2FkSWRlbnRpdHkiV9pBBG5hbWWK6jAZYmIud29ya2xvYWRJZGVudGl0aWVzLmdldJDqMAGg6jAEqOowCYLT5JMCIRIfL3YxL3tuYW1lPXdvcmtsb2FkSWRlbnRpdGllcy8qfRKJAgoWTGlzdFdvcmtsb2FkSWRlbnRpdGllcxIqLmJ5dGViYXNlLnYxLkxpc3RXb3JrbG9hZElkZW50aXRpZXNSZXF1ZXN0GisuYnl0ZWJhc2UudjEuTGlzdFdvcmtsb2FkSWRlbnRpdGllc1Jlc3BvbnNlIpUB2kEGcGFyZW50iuowGmJiLndvcmtsb2FkSWRlbnRpdGllcy5saXN0kOowAaDqMASo6jAJgtPkkwJcWi4SLC92MS97cGFyZW50PXdvcmtzcGFjZXMvKn0vd29ya2xvYWRJZGVudGl0aWVzEiovdjEve3BhcmVudD1wcm9qZWN0cy8qfS93b3JrbG9hZElkZW50aXRpZXMSggIKFlVwZGF0ZVdvcmtsb2FkSWRlbnRpdHkSKi5ieXRlYmFzZS52MS5VcGRhdGVXb3JrbG9hZElkZW50aXR5UmVxdWVzdBodLmJ5dGViYXNlLnYxLldvcmtsb2FkSWRlbnRpdHkinAHaQR13b3JrbG9hZF9pZGVudGl0eSx1cGRhdGVfbWFza4rqMBxiYi53b3JrbG9hZElkZW50aXRpZXMudXBkYXRlkOowAZjqMAGg6jADqOowBoLT5JMCRjoRd29ya2xvYWRfaWRlbnRpdHkyMS92MS97d29ya2xvYWRfaWRlbnRpdHkubmFtZT13b3JrbG9hZElkZW50aXRpZXMvKn0SvAEKFkRlbGV0ZVdvcmtsb2FkSWRlbnRpdHkSKi5ieXRlYmFzZS52MS5EZWxldGVXb3JrbG9hZElkZW50aXR5UmVxdWVzdBoWLmdvb2dsZS5wcm90b2J1Zi5FbXB0eSJe2kEEbmFtZYrqMBxiYi53b3JrbG9hZElkZW50aXRpZXMuZGVsZXRlkOowAZjqMAGg6jAEqOowCYLT5JMCISofL3YxL3tuYW1lPXdvcmtsb2FkSWRlbnRpdGllcy8qfRLOAQoYVW5kZWxldGVXb3JrbG9hZElkZW50aXR5EiwuYnl0ZWJhc2UudjEuVW5kZWxldGVXb3JrbG9hZElkZW50aXR5UmVxdWVzdBodLmJ5dGViYXNlLnYxLldvcmtsb2FkSWRlbnRpdHkiZYrqMB5iYi53b3JrbG9hZElkZW50aXRpZXMudW5kZWxldGWQ6jABmOowAaDqMASo6jAJgtPkkwItOgEqIigvdjEve25hbWU9d29ya2xvYWRJZGVudGl0aWVzLyp9OnVuZGVsZXRlQrIBCg9jb20uYnl0ZWJhc2UudjFCHFdvcmtsb2FkSWRlbnRpdHlTZXJ2aWNlUHJvdG9QAVo0Z2l0aHViLmNvbS9ieXRlYmFzZS9ieXRlYmFzZS9iYWNrZW5kL2dlbmVyYXRlZC1nby92MaICA0JYWKoCC0J5dGViYXNlLlYxygILQnl0ZWJhc2VcVjHiAhdCeXRlYmFzZVxWMVxHUEJNZXRhZGF0YeoCDEJ5dGViYXNlOjpWMWIGcHJvdG8z", [file_buf_validate_validate, file_google_api_annotations, file_google_api_client, file_google_api_field_behavior, file_google_api_resource, file_google_protobuf_empty, file_google_protobuf_field_mask, file_google_protobuf_timestamp, file_v1_annotation, file_v1_common]);
 
 /**
  * Describes the message bytebase.v1.WorkloadIdentity.
